Save and revert implemented and code refactoring
I implemented the cmd-s and cmd-r behavior and got sql table changes
caching instead of immediately saving. also simplified a lot of code
partially by making it so the model is instatiated from QML and by
preloading each tab, which I was hesitant to do at first but the
benefits seem to be far outweighing my initial resons for hesitating.