Move TaskEditView so it can access the model
TaskEditView needed to be adjacent to the Task Row list so that both
could access the same model. TaskEditView now auto saves as a result,
simplifying a bunch of code. Should be much easier to populate the
checklists in the checklist tab now, though this isn't set up just yet.